Add call to helper function in Unit Tests
On bfdprofile_controller_test.go there is helper function
called validateConfigMatchesYaml which is being used in this file.
moved this function to suite_test.go

This commit change the files:
controllers/addresspool_controller_test.go
controllers/bgppeer_controller_test.go
And make them use this helper function as well.
