+++ controllers/addresspool_controller_test.go
@@ -2,24 +2,19 @@ package controllers
 
 import (
 	"context"
-	"time"
 
 	"k8s.io/utils/pointer"
 
 	"github.com/metallb/metallb-operator/api/v1beta1"
 	metallbv1beta1 "github.com/metallb/metallb-operator/api/v1beta1"
-	"github.com/metallb/metallb-operator/pkg/apply"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/types"
 )
 
 var _ = Describe("AddressPool Controller", func() {
 	Context("Creating AddressPool object Layer2 Config", func() {
 		autoAssign := false
-		configmap := &corev1.ConfigMap{}
 
 		BeforeEach(func() {
 			metallb := &metallbv1beta1.MetalLB{
@@ -69,31 +64,19 @@ var _ = Describe("AddressPool Controller", func() {
 			Expect(err).ToNot(HaveOccurred())
 
 			By("Checking ConfigMap is created and matches test-addresspool1 configuration")
-			Eventually(func() (string, error) {
-				err := k8sClient.Get(context.Background(), types.NamespacedName{Name: apply.MetalLBConfigMap, Namespace: MetalLBTestNameSpace}, configmap)
-				if err != nil {
-					return "", err
-				}
-				return configmap.Data[apply.MetalLBConfigMap], err
-			}, 2*time.Second, 200*time.Millisecond).Should(MatchYAML(`address-pools:
+			validateConfigMatchesYaml(`address-pools:
 - name: test-addresspool1
   protocol: layer2
   auto-assign: false
   addresses:
   - 1.1.1.1-1.1.1.100
-`))
+`)
 			By("Creating 2nd AddressPool resource")
 			err = k8sClient.Create(context.Background(), addressPool2)
 			Expect(err).ToNot(HaveOccurred())
 
 			By("Checking ConfigMap is created and matches test-addresspool1 & 2 configuration")
-			Eventually(func() (string, error) {
-				err := k8sClient.Get(context.Background(), types.NamespacedName{Name: apply.MetalLBConfigMap, Namespace: MetalLBTestNameSpace}, configmap)
-				if err != nil {
-					return "", err
-				}
-				return configmap.Data[apply.MetalLBConfigMap], err
-			}, 2*time.Second, 200*time.Millisecond).Should(MatchYAML(`address-pools:
+			validateConfigMatchesYaml(`address-pools:
 - name: test-addresspool1
   protocol: layer2
   auto-assign: false
@@ -104,42 +87,31 @@ var _ = Describe("AddressPool Controller", func() {
   auto-assign: false
   addresses:
   - 2.2.2.2-2.2.2.100
-`))
+`)
 			By("Deleting 1st AddressPool resource")
 			err = k8sClient.Delete(context.Background(), addressPool1)
 			Expect(err).ToNot(HaveOccurred())
 
 			By("Checking ConfigMap is created and matches test-addresspool2 configuration")
-			Eventually(func() (string, error) {
-				err := k8sClient.Get(context.Background(), types.NamespacedName{Name: apply.MetalLBConfigMap, Namespace: MetalLBTestNameSpace}, configmap)
-				if err != nil {
-					return "", err
-				}
-				return configmap.Data[apply.MetalLBConfigMap], err
-			}, 2*time.Second, 200*time.Millisecond).Should(MatchYAML(`address-pools:
+			validateConfigMatchesYaml(`address-pools:
 - name: test-addresspool2
   protocol: layer2
   auto-assign: false
   addresses:
   - 2.2.2.2-2.2.2.100
 
-`))
+`)
 			By("Deleting 2nd AddressPool resource")
 			err = k8sClient.Delete(context.Background(), addressPool2)
 			Expect(err).ToNot(HaveOccurred())
 
 			By("Checking ConfigMap is empty")
-			Eventually(func() string {
-				err := k8sClient.Get(context.Background(), types.NamespacedName{Name: apply.MetalLBConfigMap, Namespace: MetalLBTestNameSpace}, configmap)
-				Expect(err).ToNot(HaveOccurred())
-				return configmap.Data["config"]
-			}, 2*time.Second, 200*time.Millisecond).Should(MatchYAML("{}"))
+			validateConfigMatchesYaml("{}")
 		})
 	})
 
 	Context("Creating AddressPool object BGP Config", func() {
 		autoAssign := false
-		configmap := &corev1.ConfigMap{}
 		addressPool := &v1beta1.AddressPool{
 			ObjectMeta: metav1.ObjectMeta{
 				Name:      "test-addresspool-bgp",
@@ -190,13 +162,7 @@ var _ = Describe("AddressPool Controller", func() {
 
 			// Checking ConfigMap is created
 			By("Checking ConfigMap is created and matches test-addresspool-bgp configuration")
-			Eventually(func() (string, error) {
-				err := k8sClient.Get(context.Background(), types.NamespacedName{Name: apply.MetalLBConfigMap, Namespace: MetalLBTestNameSpace}, configmap)
-				if err != nil {
-					return "", err
-				}
-				return configmap.Data[apply.MetalLBConfigMap], err
-			}, 2*time.Second, 200*time.Millisecond).Should(MatchYAML(`address-pools:
+			validateConfigMatchesYaml(`address-pools:
 - name: test-addresspool-bgp
   protocol: bgp
   auto-assign: false
@@ -210,7 +176,7 @@ var _ = Describe("AddressPool Controller", func() {
     aggregation-length: 24
     aggregation-length-v6: 124
     localpref: 100
-`))
+`)
 		})
 	})
 })
