+++ controllers/suite_test.go
@@ -22,11 +22,14 @@ import (
 	"os"
 	"path/filepath"
 	"testing"
+	"time"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/kubernetes/scheme"
 	"k8s.io/client-go/rest"
 	ctrl "sigs.k8s.io/controller-runtime"
@@ -38,6 +41,7 @@ import (
 
 	metallbv1alpha1 "github.com/metallb/metallb-operator/api/v1alpha1"
 	metallbv1beta1 "github.com/metallb/metallb-operator/api/v1beta1"
+	"github.com/metallb/metallb-operator/pkg/apply"
 	// +kubebuilder:scaffold:imports
 )
 
@@ -134,3 +138,18 @@ var _ = AfterSuite(func() {
 	err := testEnv.Stop()
 	Expect(err).ToNot(HaveOccurred())
 })
+
+func validateConfigMatchesYaml(toMatch string) {
+	configmap := &corev1.ConfigMap{}
+	EventuallyWithOffset(1, func() (string, error) {
+		err := k8sClient.Get(context.Background(),
+			types.NamespacedName{Name: apply.MetalLBConfigMap, Namespace: MetalLBTestNameSpace}, configmap)
+		if err != nil {
+			if errors.IsNotFound(err) {
+				return "", nil
+			}
+			return "", err
+		}
+		return configmap.Data[apply.MetalLBConfigMap], err
+	}, 2*time.Second, 200*time.Millisecond).Should(MatchYAML(toMatch))
+}
